+++ src/renderer/src/App/qmemory.impl.tsx
@@ -1,5 +1,5 @@
 /**
- * App - Memu Implementation
+ * App - Qmemory Implementation
  * Full-featured app with all messaging platforms
  */
 import { useState, useEffect } from 'react'
@@ -20,7 +20,7 @@ import { appIcon } from '../assets'
 type NavItem = 'telegram' | 'discord' | 'whatsapp' | 'slack' | 'line' | 'feishu' | 'settings'
 type AppNavItem = Exclude<NavItem, 'settings'>
 
-const LAST_APP_TAB_KEY = 'memu-last-app-tab'
+const LAST_APP_TAB_KEY = 'qmemory-last-app-tab'
 
 // Get saved tab or default to telegram
 function getSavedAppTab(): AppNavItem {
@@ -176,7 +176,7 @@ export function QmemoryApp(): JSX.Element {
       <div className="h-screen flex flex-col items-center justify-center bg-gradient-to-b from-[var(--bg-base)] via-[var(--bg-secondary)] to-[var(--bg-tertiary)]">
         <div className="mb-8">
           <div className="w-28 h-28 rounded-3xl bg-[var(--icon-bg)] flex items-center justify-center shadow-lg">
-            <img src={appIcon} alt="memU" className="w-24 h-24 rounded-2xl" />
+            <img src={appIcon} alt="Project 2501" className="w-24 h-24 rounded-2xl" />
           </div>
         </div>
