+++ src/renderer/src/assets/index.ts
@@ -2,10 +2,10 @@
  * Asset exports
  */
 
-import appIconMemu from './app-icon-2501.png'
+import appIcon2501 from './app-icon-2501.png'
 
 // App icon
-export const appIcon = appIconMemu
+export const appIcon = appIcon2501
 
 // Re-export other assets
 export { default as logoSvg } from './logo.svg'
