+++ src/renderer/src/components/Layout/Sidebar/index.tsx
@@ -2,7 +2,7 @@
  * Sidebar Component
  */
 import type { ComponentType } from 'react'
-import { QmemorySidebar } from './memu.impl'
+import { QmemorySidebar } from './qmemory.impl'
 import type { QmemoryNavItem, QmemorySidebarProps } from './types'
 
 // Export the Sidebar component
