+++ src/renderer/src/components/Settings/DataSettings/index.tsx
@@ -1,3 +1,3 @@
-import { QmemoryDataSettings } from './memu.impl'
+import { QmemoryDataSettings } from './qmemory.impl'
 
 export const DataSettings = QmemoryDataSettings
