+++ src/renderer/src/components/Settings/SettingsView/qmemory.impl.tsx
